requirements.txt
flask
pymysql
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
